requirements.txt
langchain
faiss-cpu
fastapi
uvicorn
pypdf
python-dotenv
google-generativeai

langchain-community
langchain-google-genai
